Move dump function to separate file
