+++ pkg/web/dump.go
@@ -1,0 +1,12 @@
+package web
+
+import "encoding/json"
+
+func Dump(data interface{}) (string, error) {
+	jsonData, err := json.Marshal(data)
+	if err != nil {
+		return "", err
+	}
+
+	return string(jsonData), nil
+}
